The issue was that checkers needed to have the exact same prototype in order to be stored in std::vector<> checkers_, making them unyieldy and hard to reuse.

Solution was to create a generic function check::exists, which allows us to check whether a std::string exists in every serv_.map.

Checkers are now way more streamlined with better code reuse.

+++ src/builder/command-validation/checkers/check_mode.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_mode.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aljulien < aljulien@student.42lyon.fr>     +#+  +:+       +#+        */
+/*   By: cdomet-d <cdomet-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/26 10:58:28 by cdomet-d          #+#    #+#             */
-/*   Updated: 2025/04/03 15:59:36 by aljulien         ###   ########.fr       */
+/*   Updated: 2025/04/04 17:57:07 by cdomet-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -99,9 +99,14 @@ bool check::mode_::formatArgs(CmdSpec &cmd) {
 	return true;
 }
 
-bool check::mode(CmdSpec &cmd, int idx) {
+bool check::mode(CmdSpec &cmd, size_t idx) {
 	(void)idx;
 	if (!check::mode_::formatArgs(cmd))
 		return false;
 	return true;
 }
+
+//TODO: check if there is a target and that the client is on it.
+//TODO: +l: block if arg is not an interger
+//TODO: +o: check client exist and is on channel
+//TODO: +k: check that 26 > len > 8
